refactor: 역할에 맞는 메소드명으로 변경 및 이동
