+++ src/test/java/lotto/domain/LottoTest.java
@@ -3,25 +3,16 @@ package lotto.domain;
 import static java.util.Arrays.asList;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import lotto.enums.Prize;
 import org.junit.Test;
 
 public class LottoTest {
 
 	@Test
-	public void 당첨번호가_존재할때_검증() {
-		Lotto lotto = new Lotto(asList(1, 2, 3, 4, 5, 6));
+	public void 로또_당첨_검증() {
+		Lotto lotto = new Lotto(asList(1, 2, 3, 4, 5, 6)); // 3개 매치
+		WinningNumber winningNumber = new WinningNumber("1, 2, 3, 7, 8, 9");
 
-		int winningNumber = 3;
-
-		assertThat(lotto.match(winningNumber)).isTrue();
-	}
-
-	@Test
-	public void 당첨번호가_존재하지않을때_검증() {
-		Lotto lotto = new Lotto(asList(1, 2, 3, 4, 5, 6));
-
-		int winningNumber = 9;
-
-		assertThat(lotto.match(winningNumber)).isFalse();
+		assertThat(lotto.match(winningNumber)).isEqualTo(Prize.MATCH3);
 	}
 }
