+++ src/test/java/lotto/domain/WinningNumberTest.java
@@ -1,18 +1,26 @@
 package lotto.domain;
 
-import static com.google.common.primitives.Ints.asList;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import lotto.enums.Prize;
 import org.junit.Test;
 
 public class WinningNumberTest {
 
 	@Test
-	public void 로또_당첨_검증() {
-		WinningNumber winningNumber = new WinningNumber("1, 2, 3, 7, 8, 9");
-		Lotto lotto = new Lotto(asList(1, 2, 3, 4, 5, 6)); // 3개 매치
+	public void 당첨번호가_존재할때_검증() {
+		WinningNumber winningNumber = new WinningNumber("1, 2, 3, 4, 5, 6");
 
-		assertThat(winningNumber.match(lotto)).isEqualTo(Prize.MATCH3);
+		int number = 3;
+
+		assertThat(winningNumber.hasNumber(number)).isTrue();
+	}
+
+	@Test
+	public void 당첨번호가_존재하지않을때_검증() {
+		WinningNumber winningNumber = new WinningNumber("1, 2, 3, 4, 5, 6");
+
+		int number = 9;
+
+		assertThat(winningNumber.hasNumber(number)).isFalse();
 	}
 }
